Start creating the app module inside main.js
