+++ index.html
@@ -241,7 +241,7 @@
         <div id="loader">
             <img src="assets/logo.svg" id="spinner">
         </div>
-        <script src="main.js"></script>
+        <script type="module" src="main.js"></script>
     </main>
 </body>
 </html>
